requirements.txt
appdirs==1.4.4
audioread==2.1.9
azure-core==1.21.1
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
decorator==4.4.2
idna==3.3
imageio==2.13.5
imageio-ffmpeg==0.4.5
isodate==0.6.1
joblib==1.1.0
jq==1.2.1
librosa==0.8.1
llvmlite==0.38.0
moviepy==1.0.3
msrest==0.6.21
numba==0.55.1
numpy==1.21.5
oauthlib==3.1.1
packaging==21.3
Pillow==9.0.0
pooch==1.6.0
proglog==0.1.9
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.0
resampy==0.2.2 s
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
SoundFile==0.10.3.post1
swagger-client @ file:///Users/hampras/Downloads/python-client
threadpoolctl==3.1.0
torch==1.10.2
tqdm==4.62.3
typing==3.7.4.3
typing_extensions==4.0.1
urllib3==1.26.8
webrtcvad==2.0.10
pytube==12.0.0
